The options are initially filtered, then again when control focused
Not sure if this is the right behaviour, but it does make much of the logic simpler.
It is also possibly easier for a control that is part of a form, where one select influences 
the filtering of another. The current implementation means the filter gets called each
time, which makes the logic of such a form much simpler.

Tests have been amended such that the control is not initially focused, but the 
`filterOption` callback is still called for each option once.