Bugfix/remove circular refs (#1855)
* Fix issue with Circular references.

Because Select was used in Async, AsyncCreatable, and Creatable and those components were ref in
Select, there was a circular dependency. It seems mostackage managers handled this ok, but this
may cause an issue with hoisting (e.g. rollup and webpack 3).# Please enter the commit message for your changes. Lines starting

* Add focus test back in to Async

* Make components static fields on Select instance